+++ WireBender3D/Assets/Prefabs/WireBenderMachine.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 7b4567e73c6455b41a2c7b72a63f130a
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
