+++ WireBender3D/Assets/Source/Collisions.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4739208c1a590bd4294fe464ccd20a27
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
