+++ WireBender3D/Assets/Source/Collisions/MachineCollisions.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2790938dc49fd3042ab73a0191ba1586
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
