+++ WireBender3D/Assets/Source/WireCreator/WireUserCreator.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d92209a90d2a21847a519f189a448040
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
